refactor(tools): move swap helpers to collections
